# openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": true,
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "FacetCount": {
        "additionalProperties": true,
        "properties": {
          "facetCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facetcategory"
          },
          "facetLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facetlabel"
          },
          "facetName": {
            "title": "Facetname",
            "type": "string"
          },
          "facetValue": {
            "title": "Facetvalue",
            "type": "string"
          },
          "scopedCount": {
            "default": 0,
            "title": "Scopedcount",
            "type": "integer"
          }
        },
        "required": [
          "facetName",
          "facetValue"
        ],
        "title": "FacetCount",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RelationEvidenceResponse": {
        "additionalProperties": true,
        "properties": {
          "annotations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Annotations",
            "type": "array"
          },
          "evidence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Evidence",
            "type": "array"
          }
        },
        "title": "RelationEvidenceResponse",
        "type": "object"
      },
      "RelationFilterOptions": {
        "additionalProperties": true,
        "properties": {
          "interactionTypes": {
            "items": {
              "type": "string"
            },
            "title": "Interactiontypes",
            "type": "array"
          },
          "predicatesByCategory": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Predicatesbycategory",
            "type": "object"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "title": "RelationFilterOptions",
        "type": "object"
      },
      "RelationPayloadsResponse": {
        "additionalProperties": true,
        "properties": {
          "payloads": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Payloads",
            "type": "array"
          },
          "relationPk": {
            "title": "Relationpk",
            "type": "string"
          }
        },
        "required": [
          "relationPk"
        ],
        "title": "RelationPayloadsResponse",
        "type": "object"
      },
      "RelationRow": {
        "additionalProperties": true,
        "properties": {
          "objectEntity": {
            "$ref": "#/components/schemas/EntitySummary"
          },
          "relation": {
            "$ref": "#/components/schemas/RelationSummary"
          },
          "subjectEntity": {
            "$ref": "#/components/schemas/EntitySummary"
          }
        },
        "required": [
          "relation",
          "subjectEntity",
          "objectEntity"
        ],
        "title": "RelationRow",
        "type": "object"
      },
      "RelationSearchRequest": {
        "additionalProperties": true,
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/SearchFilters"
          },
          "limit": {
            "default": 20,
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "default": 0,
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "resources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resources"
          }
        },
        "title": "RelationSearchRequest",
        "type": "object"
      },
      "RelationSearchResponse": {
        "additionalProperties": true,
        "properties": {
          "elapsed_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Ms"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "relations": {
            "items": {
              "$ref": "#/components/schemas/RelationSummary"
            },
            "title": "Relations",
            "type": "array"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RelationRow"
            },
            "title": "Rows",
            "type": "array"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          }
        },
        "title": "RelationSearchResponse",
        "type": "object"
      },
      "RelationSummary": {
        "additionalProperties": true,
        "properties": {
          "displayLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displaylabel"
          },
          "evidenceCount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidencecount"
          },
          "interactionClass": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interactionclass"
          },
          "isDirected": {
            "default": false,
            "title": "Isdirected",
            "type": "boolean"
          },
          "objectEntityPk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectentitypk"
          },
          "participantTypes": {
            "items": {
              "type": "string"
            },
            "title": "Participanttypes",
            "type": "array"
          },
          "predicate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Predicate"
          },
          "relationCategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Relationcategory"
          },
          "relationPk": {
            "title": "Relationpk",
            "type": "string"
          },
          "sign": {
            "default": 0,
            "title": "Sign",
            "type": "integer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "subjectEntityPk": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subjectentitypk"
          }
        },
        "required": [
          "relationPk"
        ],
        "title": "RelationSummary",
        "type": "object"
      },
...
      "ResourcesResponse": {
        "additionalProperties": true,
        "properties": {
          "resources": {
            "items": {
              "$ref": "#/components/schemas/ResourceRecord"
            },
            "title": "Resources",
            "type": "array"
          }
        },
        "title": "ResourcesResponse",
        "type": "object"
      },
...
      "ScopedRelationFacetsRequest": {
        "additionalProperties": true,
        "properties": {
          "annotationTermIds": {
            "items": {
              "type": "string"
            },
            "title": "Annotationtermids",
            "type": "array"
          },
          "causal_mechanism_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Causal Mechanism Qualifier",
            "type": "array"
          },
          "endpointMode": {
            "anyOf": [
              {
                "enum": [
                  "any",
                  "both"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "any",
            "title": "Endpointmode"
          },
          "entityIds": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Entityids",
            "type": "array"
          },
          "entity_ids": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "filters": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchFilters"
              },
              {
                "type": "null"
              }
            ]
          },
          "interactionTypes": {
            "items": {
              "type": "string"
            },
            "title": "Interactiontypes",
            "type": "array"
          },
          "mode": {
            "anyOf": [
              {
                "enum": [
                  "union",
                  "intersection"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          },
          "object_aspect_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Object Aspect Qualifier",
            "type": "array"
          },
          "object_direction_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Object Direction Qualifier",
            "type": "array"
          },
          "predicates": {
            "items": {
              "type": "string"
            },
            "title": "Predicates",
            "type": "array"
          },
          "relation_categories": {
            "items": {
              "type": "string"
            },
            "title": "Relation Categories",
            "type": "array"
          },
          "resources": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resources"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "taxonomyIds": {
            "items": {
              "type": "string"
            },
            "title": "Taxonomyids",
            "type": "array"
          },
          "taxonomyLimit": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Taxonomylimit"
          },
          "taxonomyQuery": {
            "default": "",
            "title": "Taxonomyquery",
            "type": "string"
          }
        },
        "title": "ScopedRelationFacetsRequest",
        "type": "object"
      },
      "SearchFilters": {
        "additionalProperties": true,
        "description": "One internal filter contract, accepting historical wire aliases once.",
        "properties": {
          "annotation_term_ids": {
            "items": {
              "type": "string"
            },
            "title": "Annotation Term Ids",
            "type": "array"
          },
          "causal_mechanism_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Causal Mechanism Qualifier",
            "type": "array"
          },
          "entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Entity Ids",
            "type": "array"
          },
          "entity_pks": {
            "items": {
              "type": "string"
            },
            "title": "Entity Pks",
            "type": "array"
          },
          "entity_types": {
            "items": {
              "type": "string"
            },
            "title": "Entity Types",
            "type": "array"
          },
          "gene_mode": {
            "default": "any",
            "enum": [
              "any",
              "both",
              "source",
              "source_only",
              "target",
              "target_only"
            ],
            "title": "Gene Mode",
            "type": "string"
          },
          "interaction_types": {
            "items": {
              "type": "string"
            },
            "title": "Interaction Types",
            "type": "array"
          },
          "is_directed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Directed"
          },
          "min_evidence_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Evidence Count"
          },
          "object_aspect_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Object Aspect Qualifier",
            "type": "array"
          },
          "object_direction_qualifier": {
            "items": {
              "type": "string"
            },
            "title": "Object Direction Qualifier",
            "type": "array"
          },
          "predicates": {
            "items": {
              "type": "string"
            },
            "title": "Predicates",
            "type": "array"
          },
          "relation_categories": {
            "items": {
              "type": "string"
            },
            "title": "Relation Categories",
            "type": "array"
          },
          "scope_endpoint_mode": {
            "default": "any",
            "enum": [
              "any",
              "both",
              "source",
              "source_only",
              "target",
              "target_only"
            ],
            "title": "Scope Endpoint Mode",
            "type": "string"
          },
          "scope_entity_ids": {
            "items": {
              "type": "string"
            },
            "title": "Scope Entity Ids",
            "type": "array"
          },
          "sign": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sign"
          },
          "signs": {
            "items": {
              "type": "integer"
            },
            "title": "Signs",
            "type": "array"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          },
          "taxonomy_ids": {
            "items": {
              "type": "string"
            },
            "title": "Taxonomy Ids",
            "type": "array"
          }
        },
        "title": "SearchFilters",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "REST API for querying OmniPath knowledge graphs directly from nested Parquet files via DuckDB. The Svelte explorer uses `/api/*` (and `/app-api/*` as an application alias). Interactive docs: `/api/docs`.",
    "title": "OmniPath Parquet API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/relations/filter-options": {
      "get": {
        "description": "Predicate/source/type enumerations for the relation explorer.",
        "operationId": "relation_filter_options_relations_filter_options_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationFilterOptions"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Relation Filter Options",
        "tags": [
          "relations"
        ]
      }
    },
    "/relations/scoped-facets": {
      "post": {
        "description": "Facet counts for the current relation selection.",
        "operationId": "scoped_relation_facets_relations_scoped_facets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScopedRelationFacetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FacetCount"
                  },
                  "title": "Response Scoped Relation Facets Relations Scoped Facets Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Scoped Relation Facets",
        "tags": [
          "relations"
        ]
      }
    },
    "/relations/search": {
      "post": {
        "description": "Filtered relation search with offset pagination.",
        "operationId": "search_relations_relations_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RelationSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Search Relations",
        "tags": [
          "relations"
        ]
      }
    },
    "/relations/{relation_pk}": {
      "get": {
        "description": "Fetch one relation plus hydrated endpoints.",
        "operationId": "get_relation_relations__relation_pk__get",
        "parameters": [
          {
            "in": "path",
            "name": "relation_pk",
            "required": true,
            "schema": {
              "title": "Relation Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Relation Relations  Relation Pk  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Get Relation",
        "tags": [
          "relations"
        ]
      }
    },
    "/relations/{relation_pk}/evidence": {
      "get": {
        "description": "Nested evidence and annotations for one relation.",
        "operationId": "relation_evidence_relations__relation_pk__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "relation_pk",
            "required": true,
            "schema": {
              "title": "Relation Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationEvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Relation Evidence",
        "tags": [
          "relations"
        ]
      }
    },
    "/relations/{relation_pk}/payloads": {
      "get": {
        "description": "Raw source payloads attached to a relation.",
        "operationId": "relation_payloads_relations__relation_pk__payloads_get",
        "parameters": [
          {
            "in": "path",
            "name": "relation_pk",
            "required": true,
            "schema": {
              "title": "Relation Pk",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RelationPayloadsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Relation Payloads",
        "tags": [
          "relations"
        ]
      }
    },
...
    "/resources": {
      "get": {
        "description": "List indexed resources. `shape=svelte` returns the explorer catalog.",
        "operationId": "list_resources_resources_get",
        "parameters": [
          {
            "description": "Pass `svelte` for the explorer catalog shape.",
            "in": "query",
            "name": "shape",
            "required": false,
            "schema": {
              "default": "web",
              "description": "Pass `svelte` for the explorer catalog shape.",
              "title": "Shape",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "List Resources",
        "tags": [
          "resources"
        ]
      }
    },
...
  },
...
}